requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
black==25.11.0
certifi==2026.4.22
cffi==2.0.0
cfgv==3.4.0
click==8.1.8
cryptography==47.0.0
distlib==0.4.0
dnspython==2.7.0
ecdsa==0.19.2
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.111.0
fastapi-cli==0.0.22
filelock==3.19.1
flake8==7.3.0
greenlet==3.2.5
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.15
idna==3.13
importlib_metadata==8.7.1
isort==6.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.10.0
orjson==3.11.5
packaging==26.1
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.4.0
pre_commit==4.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.3
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.7.1
pydantic-settings==2.11.0
pydantic_core==2.18.2
pyflakes==3.4.0
Pygments==2.20.0
python-discovery==1.2.2
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.4.1
PyYAML==6.0.3
rich==15.0.0
rich-toolkit==0.19.7
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.30
starlette==0.37.2
tomli==2.4.1
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
uvicorn==0.30.1
uvloop==0.22.1
virtualenv==21.2.4
watchfiles==1.1.1
websockets==15.0.1
zipp==3.23.1
